Fix failing fuzz

+++ src/test/DeltaNeutralLp.t.sol
@@ -198,7 +198,7 @@ contract DeltaNeutralTest is TestPlus {
         assertEq(sushiBalance, 0);
     }
 
-    function testTVLFuzz(uint256 assets) public {
+    function testTVLFuzz(uint64 assets) public {
         // Max borrowable WETH available in AAVE in this block is around 1334.66 WETH or 2178919.22 USDC.
         // So technically we should be able to take position with around 2178919.22 / ((4 / 7) * (3 / 4)) = 5084144.84 USDC
         vm.assume(assets < 4e12 && assets > 1e5);
